refactor logic that updates a service entry with client id returned from the server
